requirements.txt
RPi.GPIO==0.7.0
schedule==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==2.3.6
blinker==1.6.2
click==8.1.6
flask==2.3.2
importlib-metadata==6.8.0
itsdangerous==2.1.2
zipp==3.16.2